+++ backend/pet/src/main/java/com/sns/pet/service/PetService.java
@@ -8,7 +8,7 @@ public interface PetService {
     // 모든 반려동물 목록 조회
     public List<PetDto> findAllPets(Long userNumber) throws Exception;
     // 특정 반려동물 상세 조회
-    public PetDto findPet(Long userNumber, int petNumber) throws Exception;
+    public PetDto findPet(Long userNumber, Long petNumber) throws Exception;
     // 유저의 반려동물 이름 중복 체크
     public boolean checkPetName(Long userNumber, String petName) throws Exception;
     // 반려동물 정보 저장
@@ -16,5 +16,5 @@ public interface PetService {
     // 반려동물 정보 수정
     public boolean modifyPet(PetDto petDto) throws Exception;
     // 반려동물 정보 삭제
-    public boolean removePet(Long userNumber, int petNumber) throws Exception;
+    public boolean removePet(Long userNumber, Long petNumber) throws Exception;
 }
